[PHPStanRules] Move ForbiddenMethodOrStaticCallInForeachRule to ForbiddenComplexForeachIfExprRule
