+++ packages/phpstan-rules/src/Rules/Complexity/ForbiddenAssignInLoopRule.php
