+++ packages/phpstan-rules/tests/Rules/Complexity/ForbiddenComplexForeachIfExprRule/Fixture/WithMethodCallWithParameterNotFromThis.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Symplify\PHPStanRules\Tests\Rules\ForbiddenMethodOrStaticCallInForeachRule\Fixture;
+namespace Symplify\PHPStanRules\Tests\Rules\Complexity\ForbiddenComplexForeachIfExprRule\Fixture;
 
 class WithMethodCallWithParameterNotFromThis
 {
