+++ packages/phpstan-rules/tests/Rules/Complexity/ForbiddenComplexForeachIfExprRule/ForbiddenMethodOrStaticCallInForeachRuleTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Symplify\PHPStanRules\Tests\Rules\ForbiddenMethodOrStaticCallInForeachRule;
+namespace Symplify\PHPStanRules\Tests\Rules\ForbiddenComplexForeachIfExprRule;
 
 use Iterator;
 use PHPStan\Rules\Rule;
