+++ packages/phpstan-rules/tests/Rules/Complexity/ForbiddenComplexForeachIfExprRule/config/configured_rule.neon
