--- packages/phpstan-rules/tests/Rules/ForbiddenMethodOrStaticCallInForeachRule/config/configured_rule.neon
@@ -1,3 +1,0 @@
-includes:
-    - ../../../../config/symplify-rules.neon
-    - ../../../../config/services/services.neon
